update migration file

+++ prisma/migrations/20231126094603_init_migration/migration.sql
@@ -15,7 +15,7 @@ CREATE TABLE "User" (
 CREATE TABLE "Product" (
     "id" SERIAL NOT NULL,
     "name" TEXT NOT NULL,
-    "price" BIGINT NOT NULL,
+    "price" INTEGER NOT NULL,
     "description" TEXT NOT NULL,
     "categoryId" INTEGER NOT NULL,
